Postpone action

+++ src/main/java/lv/sit/todo/db/Item.java
@@ -23,5 +23,5 @@ public class Item
     public int delete;
 
     @ColumnInfo(name = "postpone")
-    public int postpone;
+    public long postpone;
 }
